Modularise the jupyter notebook into structured project folders and modules.
