+++ archive/ci_build/flake8
@@ -1,0 +1,16 @@
+[flake8]
+
+ignore =
+    E111,
+    E203,
+    E225,
+    E302, 
+    E305,
+    E501, 
+    E731, 
+    W191,
+    W291,
+    W293,
+    W501,
+    W503, 
+    W504
