--- chatbot_attention_mechanism/README.md
@@ -1,68 +1,0 @@
-**AI Chatbot with Attention Mechanism**
-
-**Project Overview**
-
-This project focuses on developing an AI chatbot that leverages advanced Natural Language Processing (NLP) techniques, including tokenization, vectorization, and a self-attention mechanism, to enhance its conversational capabilities. The chatbot is designed for seamless website integration, providing users with accurate and contextually relevant responses across a wide range of topics. The inclusion of the attention mechanism allows the chatbot to focus on relevant parts of the input text, improving its understanding of context and generating better responses.
-
-**Key Features**
-
-Self-Attention Mechanism: Improves context understanding by focusing on important parts of the input text.
-Tokenization and Vectorization: Breaks down text into tokens and transforms them into numerical vectors for machine learning models.
-Dataset Integration: Merged multiple datasets into one comprehensive dataset to enhance the chatbot's range and accuracy.
-Robust Performance: Optimized code for better execution speed and improved response quality.
-Website Integration: Designed to be easily integrated into web platforms for real-time user interaction.
-
-**Technologies Used**
-
-Python: Core language for developing the chatbot.
-TensorFlow/Keras: For building and training the machine learning models.
-Scikit-learn: For data preprocessing and model evaluation.
-Pandas/Numpy: For data manipulation and analysis.
-Matplotlib: For data visualization (if needed).
-SQL: For data storage and management of user interactions.
-
-**How It Works**
-Dataset Preparation: All datasets were merged into a single comprehensive dataset covering diverse topics. This ensures that the chatbot has a wide knowledge base.
-Text Preprocessing: Tokenization and vectorization are used to convert user input into numerical data the model can understand.
-Self-Attention Mechanism: This advanced algorithm helps the chatbot to focus on the most relevant parts of the user input, improving the quality of responses.
-Model Training: The chatbot is trained on the prepared dataset using machine learning algorithms to improve its conversational abilities.
-Code Optimization: The project includes optimized code for better readability, performance, and ease of future updates.
-
-**Installation**
-
-Clone the repository:
-
-
-git clone https://github.com/redback-operations/redback-chatbot/chatbot-with-attention.git
-
-Install the necessary dependencies:
-
-pip install -r requirements.txt
-
-Run the chatbot:
-
-python run_chat.py
-
-Usage
-
-Once the chatbot is trained and running, it can be integrated into a website or used for testing via a terminal or GUI interface. Users can interact with the chatbot by asking questions, and it will respond based on the training data, utilizing the attention mechanism for more accurate and contextually relevant answers.
-
-Project Structure
-│
-├── /models/
-│   └── language_model.py       
-│
-├── /utils/
-│   └── data_loader.py          
-│
-├── /training/
-│   └── train_model.py          
-│
-├── chatbot.py                  
-├── main.py                                
-└── README.md              # Project documentation
-
-Future Improvements
-Multilingual Support: Expanding the chatbot's capabilities to support multiple languages.
-Further Dataset Expansion: Incorporating additional datasets to improve response diversity.
-Real-time Learning: Enabling the chatbot to learn from new user inputs over time for continuous improvement.
