--- rag/requirements.txt
@@ -1,7 +1,0 @@
-pymupdf4llm
-sentence-transformers
-qdrant-client
-langchain
-langchain_groq
-langchain-core
-flashrank
